+++ CHANGELOG.md
@@ -6,16 +6,18 @@
 
 * `DerivePlutusType` and everything related
 * `Plutarch.DataRepr`
+* `DeriveAsTag`, use `DeriveTagPlutusType` and `DeriveTagPLiftable` instead
 
 ## Added
 
 * `DeriveAsDataStruct``DeriveAsDataRec`, `DeriveAsSOPStruct`, `DeriveAsSOPRec` to `Plutarch.Prelude`
 * `pmatchListN`, `pmatchList`, `pmatchListUnsafe` for matching on list efficiently and ergonomically.
 * Added `ConstantHoist` option to the compilation option to allow configuring how threshold for constant to be hoisted.
+* `DeriveTagPlutusType` and `DeriveTagPLiftable`
 
 ## Changed
 
-* `plutus-core` version bound to `>=1.40.0.0 && <1.41`
+* `plutus-core` version bound to `>=1.45.0.0 && <1.46`
 * Now, default for constant hoisting is 20 bytes. This means any constant larger than 20 bytes in UPLC flat will get hoisted.
   This can be changed via compilation config.